requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
rq==1.15.1
redis==4.5.5
requests==2.31.0
pyannote.audio==3.1.1
torch==2.0.1
torchaudio==2.0.2
ffmpeg-python==0.2.0
google-cloud-speech==2.20.0
google-cloud-storage==2.9.0
firebase-admin==6.2.0
gunicorn==21.2.0
pydub==0.25.1
